papermonk scraper for plosone.org
This is a first version of a downloader module for papermonk, meant to
serve as an example for implementing other modules.

At the moment, the actual paper downloading code is not implemented
because it's silly to stuff everything into the "download" function,
maybe this concept should be split up into multiple methods instead.

version bump to: 0.0.1

(initial commit)